+++ CHANGELOG.md
@@ -20,8 +20,11 @@
 * network: new `quality` tag (Wi-Fi only).
 * Read alternative config from pipes and FIFOs (e.g. `--config
   /dev/stdin`) ([#340][340]).
+* Added `overlay` and `background` as possible `layer` values
+  ([#372][372]).
 
 [340]: https://codeberg.org/dnkl/yambar/pulls/340
+[372]: https://codeberg.org/dnkl/yambar/issues/372
 
 
 ### Changed